requirements.txt
contourpy==1.1.0
cycler==0.11.0
docutils==0.20.1
ffmpeg-python==0.2.0
filelock==3.12.4
fonttools==4.42.1
future==0.18.3
importlib-resources==6.0.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.3
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
opencv-python==4.8.0.76
packaging==23.1
Pillow==10.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
torch==2.0.1
torchimize==0.0.14
typing_extensions==4.7.1
zipp==3.16.2
